Added advancements to kits & features
